% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_values_proportions.R
\name{plot_missing_values_proportions}
\alias{plot_missing_values_proportions}
\title{Visualization of missing values}
\usage{
plot_missing_values_proportions(
  df,
  vars = tidyselect::everything(),
  point_color = "blue",
  line_color = "blue",
  ylab = "Variables",
  xlab = "Proportion of missing values"
)
}
\arguments{
\item{df}{a data frame.}

\item{vars}{a vector containing selected name(s) of variable(s) in \code{df}. These selected variables will be included in the plot. The default is to show the proportion of missing values for all variables in \code{df}.}

\item{point_color}{the color to be used for the points. The default is "blue".}

\item{line_color}{the color to be used for the lines. The default is "blue".}

\item{ylab}{the title of the y axis.}

\item{xlab}{the title of the x axis.}
}
\value{
a ggplot object visualizing the proportion of missing values for all or selected columns.
}
\description{
Makes a ggplot to visualize the proportion of missing values for all or selected columns in a data frame.
}
\examples{
# Create the plot for all variables in airquality (a dataset in Base R)
plot_missing_values_proportions(df=airquality, point_color = 'darkgreen', line_color = 'darkgreen')

# Create the plot for selected variables in airquality
plot_missing_values_proportions(df=airquality, vars = c(Ozone, Wind, Temp))

# Change the title of the y axis.
plot_missing_values_proportions(df=airquality, ylab = "Columns")
}
